+++ lib/i18n/types.ts
@@ -1439,7 +1439,9 @@ export type Messages = {
   "universe.rule.PREFERRED_NAME": string;
   "universe.rule.DATA_QUALITY": string;
   "universe.rule.LOW_TURNOVER": string;
+  "universe.rule.MANUAL_INCLUDE_WATCHLIST": string;
   "universe.updated_label": string;
+  "universe.watchlist_note": string;
   "universe.override_warning": string;
 };